# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP proxy с управлением доступом по IP (whitelist/blacklist/graylist), rate limiting (RPS/RPM/RPH/RPD/bandwidth/соединения) и мониторингом в реальном времени.",
        "title": "Proxy API",
        "contact": {
            "name": "API Support",
            "email": "artiomk00o@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/ratelimit": {
            "get": {
                "description": "GET — возвращает все правила. POST — создаёт новое правило для IP или подсети.\n\nПоля правила (все лимиты -1 = отключён):\n- `target`: \"ip\" или \"subnet\"\n- `value`: конкретный IP или CIDR (например 10.0.0.0/8)\n- `max_rps` / `max_rpm` / `max_rph` / `max_rpd`: лимиты запросов\n- `max_upload_bps` / `max_download_bps`: лимиты пропускной способности (байт/с)\n- `max_traffic_day`: дневной лимит трафика (байт)\n- `max_concurrent`: макс. одновременных соединений\n- `max_new_per_second`: макс. новых соединений в секунду",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rate Limiting"
                ],
                "summary": "Управление правилами rate limiting",
                "parameters": [
                    {
                        "description": "Правило rate limiting (только для POST)",
                        "name": "rule",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/domain.RateRule"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список всех правил",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.RateRule"
                            }
                        }
                    },
                    "201": {
                        "description": "Правило создано: {\\\"status\\\":\\\"ok\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный JSON или отсутствует поле value/target",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Метод не поддерживается",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "GET — возвращает все правила. POST — создаёт новое правило для IP или подсети.\n\nПоля правила (все лимиты -1 = отключён):\n- `target`: \"ip\" или \"subnet\"\n- `value`: конкретный IP или CIDR (например 10.0.0.0/8)\n- `max_rps` / `max_rpm` / `max_rph` / `max_rpd`: лимиты запросов\n- `max_upload_bps` / `max_download_bps`: лимиты пропускной способности (байт/с)\n- `max_traffic_day`: дневной лимит трафика (байт)\n- `max_concurrent`: макс. одновременных соединений\n- `max_new_per_second`: макс. новых соединений в секунду",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rate Limiting"
                ],
                "summary": "Управление правилами rate limiting",
                "parameters": [
                    {
                        "description": "Правило rate limiting (только для POST)",
                        "name": "rule",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/domain.RateRule"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список всех правил",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.RateRule"
                            }
                        }
                    },
                    "201": {
                        "description": "Правило создано: {\\\"status\\\":\\\"ok\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный JSON или отсутствует поле value/target",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Метод не поддерживается",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ratelimit/{id}": {
            "delete": {
                "description": "Удаляет правило по его UUID. ID правила берётся из пути: /ratelimit/{id}.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rate Limiting"
                ],
                "summary": "Удалить правило rate limiting",
                "parameters": [
                    {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "description": "UUID правила",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Правило удалено: {\\\"status\\\":\\\"deleted\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "ID правила не указан",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Метод не поддерживается",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка удаления",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "domain.RateRule": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "max_concurrent": {
                    "description": "Connection limits (-1 = disabled)",
                    "type": "integer",
                    "example": 100
                },
                "max_download_bps": {
                    "type": "integer",
                    "example": 1048576
                },
                "max_new_per_second": {
                    "type": "integer",
                    "example": 20
                },
                "max_rpd": {
                    "type": "integer",
                    "example": 100000
                },
                "max_rph": {
                    "type": "integer",
                    "example": 10000
                },
                "max_rpm": {
                    "type": "integer",
                    "example": 1000
                },
                "max_rps": {
                    "description": "Request rate limits (-1 = disabled)",
                    "type": "integer",
                    "example": 100
                },
                "max_traffic_day": {
                    "description": "Daily total traffic limit in bytes (-1 = disabled)",
                    "type": "integer",
                    "example": 104857600
                },
                "max_upload_bps": {
                    "description": "Bandwidth limits in bytes/s (-1 = disabled)",
                    "type": "integer",
                    "example": 1048576
                },
                "target": {
                    "enum": [
                        "ip",
                        "subnet"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.RuleTarget"
                        }
                    ],
                    "example": "ip"
                },
                "value": {
                    "type": "string",
                    "example": "192.168.1.1"
                }
            }
        },
        "domain.RuleTarget": {
            "type": "string",
            "enum": [
                "ip",
                "subnet"
            ],
            "x-enum-varnames": [
                "TargetIP",
                "TargetSubnet"
            ]
        },
...
    }
}